requirements.txt
discord
aiohttp
google-generativeai
deep-translator
langid